:bug: 修复Bug，满足需求

+++ app/src/main/java/com/example/smartmatch/logic/network/model/ScenarioResponse.kt
@@ -14,7 +14,7 @@ data class ScenarioResponse(
 
 data class Scenario(
     val id: Int,
-    val percentage: Int
+    val percentage: Float
 )
 data class Checkyulan(
     val if_check:Int
